Update create-comit-app npm package
