+++ create/npm/README.md
@@ -1,0 +1,5 @@
+# Create Comit App
+
+Create a new COMIT app with one command. 
+
+See https://github.com/comit-network/create-comit-app for instructions.
