+++ create/npm/package.json
@@ -1,13 +1,12 @@
 {
   "name": "create-comit-app",
   "version": "0.7.0",
-  "description": " Set up a local development environment for COMIT apps with one command.",
+  "description": "Create a new COMIT app with one command.",
   "main": "cca.js",
   "scripts": {
     "fix": "prettier --write '**/*.{ts,js,json,yml}'",
     "test": "prettier --check '**/*.{ts,js,json,yml}'",
-    "new": "./new.js",
-    "scripts": "./scripts.js"
+    "start": "./main.js"
   },
   "repository": {
     "type": "git",
@@ -26,8 +25,7 @@
   },
   "homepage": "https://github.com/comit-network/create-comit-app#readme",
   "bin": {
-    "create-comit-app": "new.js",
-    "comit-scripts": "scripts.js"
+    "create-comit-app": "main.js"
   },
   "devDependencies": {
     "prettier": "^1.18.2"
